Moves various things around to their proper locations
